Amended SetupDataLoader to include Recipe and it's components. Set up new queries in RecipeRepository and it's components. Set up RecipeController to query appropriate data from Recipe and it's components to add to the model. Set up examplerecipe.html to display the new model information.
TODO: Some of the information within RecipeCard is now redundant. RecipeCard's ID could be embeddable and borrow from Recipe. RecipeCard does not need to contain the user data or the name. Change the models to reflect this, amend examplerecipe.html to still properly display all of the data without directly pulling the RecipeCard. Set up RecipeController and add.html to intake information from Recipe and properly store it with the changes to RecipeCard.