Set up Recipe to share the same id with RecipeCard via MapsID.
TODO: RecipeCard does not need to contain the user data or the name any longer. Change the models to reflect this, amend examplerecipe.html to still properly display all of the data without directly pulling the RecipeCard. Set up RecipeController and add.html to intake information from Recipe and properly store it with the changes to RecipeCard.